城市等级浮点化起步：贴图适配几何工具 + 14种密铺枚举
- City::level 改 double；等级采样/放置回退/首都注册支持最小等级非1
- 新增 CityIconFitter：任意密铺下贴图不越界且尽量大的缩放/平移计算
- TilingType 扩展 7 半正 + 7 Laves（不做的 3.3.3.4.4/对偶五边形不预留编号）
- Config 城市形状表 levels 改 double，新增 city.tilings 通用段
- lwmap tilingByte 改为直接存 TilingType 枚举值（兼容旧 1/2）
- CityRenderer 按实数等级四舍五入取贴图（10+ 暂用9级）
- 开发思路.txt 与密铺开发备忘补充规范与代码位置表

+++ src/replay/Snapshot.cpp
@@ -242,7 +242,7 @@ bool Snapshot::deserialize(Simulation& sim, const std::string& json, std::string
             City c;
             c.id = cj[0].get<int>();
             c.ownerId = cj[1].get<int>();
-            c.level = cj[2].get<int>();
+            c.level = cj[2].get<double>();
             c.baseX = cj[3].get<int>();
             c.baseY = cj[4].get<int>();
             c.w = cj[5].get<int>();
